+++ LogBins/Base/BucketAddress.cs
@@ -7,7 +7,7 @@ namespace LogBins.Base
     public struct BucketAddress
     {
         public ushort TrainId { get; set; }
-        public int BagId { get; set; }
+        public ushort BagId { get; set; }
         public int BucketId { get; set; }
     }
 }
